feat: implement timetable wrtie
