+++ lib/core/routing/route_paths.dart
@@ -17,6 +17,7 @@ abstract class RoutePaths {
   static const String timetable = '/timetable';
   static const String timetableWrite = '/timetableWrite';
   static const String timetableList = '/timetableList';
+  static const String timetableLectureWrite = '/timetableLectureWrite';
   static const String calendar = '/calendar';
   static const String calendarDetail = '/calendarDetail';
   static const String recruitWrite = '/recruitWrite';
