+++ lib/domain/timetable/model/lecture_AI.dart
@@ -1,0 +1,3 @@
+class LectureAi {
+
+}
