+++ lib/domain/timetable/model/lecture_request.dart
@@ -1,30 +1,31 @@
 import 'package:dongsoop/domain/timetable/enum/week_day.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-part 'timetable.freezed.dart';
-part 'timetable.g.dart';
+part 'lecture_request.freezed.dart';
+part 'lecture_request.g.dart';
 
 @freezed
 @JsonSerializable()
-class Timetable with _$Timetable {
-  final int id;
+class LectureRequest with _$LectureRequest {
   final String name;
   final String professor;
   final String location;
   final WeekDay week;
   final String startAt;
   final String endAt;
+  final int year;
+  final String semester;
 
-  const Timetable({
-    required this.id,
+  const LectureRequest({
     required this.name,
     required this.professor,
     required this.location,
     required this.week,
     required this.startAt,
     required this.endAt,
+    required this.year,
+    required this.semester
   });
 
-  factory Timetable.fromJson(Map<String, dynamic> json) => _$TimetableFromJson(json);
-  Map<String, dynamic> toJson() => _$TimetableToJson(this);
+  Map<String, dynamic> toJson() => _$LectureRequestToJson(this);
 }
