--- lib/domain/timetable/model/timetable_AI.dart
@@ -1,3 +1,0 @@
-class TimetableAi {
-
-}
