+++ lib/domain/timetable/use_case/create_lecture_use_case.dart
@@ -1,0 +1,12 @@
+import 'package:dongsoop/domain/timetable/model/lecture_request.dart';
+import 'package:dongsoop/domain/timetable/repository/timetable_repository.dart';
+
+class CreateLectureUseCase {
+  final TimetableRepository _timetableRepository;
+
+  CreateLectureUseCase(this._timetableRepository,);
+
+  Future<bool> execute(LectureRequest request) async {
+    return await _timetableRepository.createLecture(request);
+  }
+}
