--- lib/domain/timetable/use_case/update_timetable_use_case.dart
@@ -1,12 +1,0 @@
-import 'package:dongsoop/domain/timetable/model/timetable.dart';
-import 'package:dongsoop/domain/timetable/repository/timetable_repository.dart';
-
-class UpdateTimetableUseCase {
-  final TimetableRepository _timetableRepository;
-
-  UpdateTimetableUseCase(this._timetableRepository,);
-
-  Future<bool> execute(Timetable timetable) async {
-    return await _timetableRepository.updateTimetable(timetable);
-  }
-}
